+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**1.0.0.2031** *(TBD)*
+
+- Changed the way the address elements were indexed. The old way did not work well for XML.
+- Updated the images used for the README.
+- Added the ability to get all the address elements in an associative array.
+
 **1.0.0.2030** *(June 11, 2018)*
 
 - Improved documentation.